api/requirements.txt
certifi==2021.5.30
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.14
django-rest-swagger==2.2.0
djangorestframework==3.9.1
idna==2.8
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==6.2.1
psycopg2==2.8.4
Pygments==2.9.0
pytz==2021.1
requests==2.22.0
simplejson==3.17.3
uritemplate==3.0.1
urllib3==1.25.7
